+++ apps/minimal/src/index.tsx
@@ -1,0 +1,7 @@
+import "expo-router/entry";
+import { Platform } from "react-native";
+
+if (Platform.OS === "web") {
+  // Keep polyfills at the top
+  localStorage.debug = "expo-audio-stream:*";
+}
