+++ apps/playground/README.md
@@ -1,0 +1,8 @@
+# Audio Playground
+
+Playground for cross platform audio streaming and feature extractions using **@siteed/expo-audio-stream**.
+
+## Try it right away
+
+https://deeeed.github.io/expo-audio-stream/
+
