+++ packages/expo-audio-stream/src/ExpoAudioStream.types.ts
@@ -1,3 +1,9 @@
+import {
+  AudioAnalysisData,
+  AudioDataEvent,
+  AudioFeaturesOptions,
+} from "./AudioAnalysis/AudioAnalysis.types";
+
 export interface AudioEventPayload {
   encoded?: string;
   buffer?: ArrayBuffer;
@@ -29,89 +35,6 @@ export interface StartAudioStreamResult {
   sampleRate?: number;
 }
 
-export interface AudioStreamStatus {
-  isRecording: boolean;
-  isPaused: boolean;
-  durationMs: number;
-  size: number;
-  interval: number;
-  mimeType: string;
-}
-
-export interface AudioDataEvent {
-  data: string | ArrayBuffer;
-  position: number;
-  fileUri: string;
-  eventDataSize: number;
-  totalSize: number;
-}
-
-export interface AudioFeatures {
-  energy: number;
-  mfcc: number[];
-  rms: number;
-  minAmplitude: number;
-  maxAmplitude: number;
-  zcr: number;
-  spectralCentroid: number;
-  spectralFlatness: number;
-  spectralRolloff: number;
-  spectralBandwidth: number;
-  chromagram: number[];
-  tempo: number;
-  hnr: number;
-}
-
-export interface AudioFeaturesOptions {
-  energy?: boolean;
-  mfcc?: boolean;
-  rms?: boolean;
-  zcr?: boolean;
-  spectralCentroid?: boolean;
-  spectralFlatness?: boolean;
-  spectralRolloff?: boolean;
-  spectralBandwidth?: boolean;
-  chromagram?: boolean;
-  tempo?: boolean;
-  hnr?: boolean;
-}
-
-export interface DataPoint {
-  id: number;
-  amplitude: number;
-  activeSpeech?: boolean;
-  dB?: number;
-  silent?: boolean;
-  features?: AudioFeatures;
-  startTime?: number;
-  endTime?: number;
-  // start / end position in bytes
-  startPosition?: number;
-  endPosition?: number;
-  // number of audio samples for this point (samples size depends on bit depth)
-  samples?: number;
-  // Id of the speaker for this point
-  speaker?: number;
-}
-
-export interface AudioAnalysisData {
-  pointsPerSecond: number; // How many consolidated value per second
-  durationMs: number; // Duration of the audio in milliseconds
-  bitDepth: number; // Bit depth of the audio
-  samples: number; // Size of the audio in bytes
-  numberOfChannels: number; // Number of audio channels
-  sampleRate: number; // Sample rate of the audio
-  dataPoints: DataPoint[];
-  amplitudeRange: {
-    min: number;
-    max: number;
-  };
-  speakerChanges?: {
-    timestamp: number;
-    speaker: number;
-  }[];
-}
-
 export type EncodingType = "pcm_32bit" | "pcm_16bit" | "pcm_8bit";
 export type SampleRate = 16000 | 44100 | 48000;
 export interface RecordingConfig {
@@ -129,7 +52,6 @@ export interface RecordingConfig {
   features?: AudioFeaturesOptions; // Feature options to extract (default is empty)
 
   // Optional paramters from web
-
   onAudioStream?: (_: AudioDataEvent) => Promise<void>; // Callback function to handle audio stream
   onProcessingResult?: (_: AudioAnalysisData) => Promise<void>; // Callback function to handle processing results
 }
