+++ packages/expo-audio-stream/src/constants.ts
@@ -1,0 +1,4 @@
+import { Platform } from "react-native";
+
+export const namespace = "expo-audio-stream";
+export const isWeb = Platform.OS === "web";
